Restructure `hyped_io` to make it board agnostic
