+++ lib/io/src/gpio.rs
@@ -1,5 +1,3 @@
-use embassy_stm32::gpio::{Input, Pin};
-
 /**
  * This trait is used to abstract the GPIO pin so that sensors can be tested with a mock GPIO pin.
  */
@@ -7,21 +5,33 @@ pub trait GpioPin {
     fn is_high(&mut self) -> bool;
 }
 
-/**
- * This struct is used to represent a physical GPIO pin on the STM32 microcontroller.
- */
-pub struct EmbassyGpio<P: Pin> {
-    pin: Input<'static, P>,
-}
+pub mod mock_gpio {
+    use heapless::Vec;
 
-impl<P: Pin> GpioPin for EmbassyGpio<P> {
-    fn is_high(&mut self) -> bool {
-        self.pin.is_high()
+    /**
+     * This struct is used to represent a mock GPIO pin that can be used for testing.
+     */
+    pub struct MockGpio {
+        current_value: bool,
+        next_values: Vec<bool, 10>,
     }
-}
 
-impl<P: Pin> EmbassyGpio<P> {
-    pub fn new(pin: Input<'static, P>) -> Self {
-        Self { pin }
+    impl crate::gpio::GpioPin for MockGpio {
+        fn is_high(&mut self) -> bool {
+            let next_value = self.next_values.pop().unwrap_or(self.current_value);
+            self.current_value = next_value;
+            self.current_value
+        }
+    }
+
+    impl MockGpio {
+        pub fn new(values: Vec<bool, 10>) -> MockGpio {
+            let mut next_values = values.clone();
+            next_values.reverse();
+            MockGpio {
+                current_value: false,
+                next_values,
+            }
+        }
     }
 }
