+++ lib/io/src/lib.rs
@@ -1,4 +1,3 @@
 #![no_std]
 
 pub mod gpio;
-pub mod mock_gpio;
